pyproject.toml
[tool.poetry]
name = "staticforms"
version = "0.1.0"
description = ""
authors = ["Edd Salkield <edd@salkield.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
email-validator = "^1.1.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
uvicorn = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
